requirements.txt
streamlit
yfinance
ta
scikit-learn
talib-binary==0.4.0.1
plotly
pandas
numpy